requirements.txt
annotated-types==0.7.0
anyio==4.9.0
dnspython==2.7.0
fastapi==0.115.11
graphql-core==3.2.6
idna==3.10
motor==3.7.0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
strawberry-graphql==0.262.5
typing_extensions==4.12.2
uvicorn
python-dotenv>=0.19.0
